+++ api/apitest/package.go
@@ -12,17 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package testtrace
-
-import (
-	"time"
-
-	"go.opentelemetry.io/otel/api/kv"
-)
-
-// Event encapsulates the properties of calls to AddEvent or AddEventWithTimestamp.
-type Event struct {
-	Timestamp  time.Time
-	Name       string
-	Attributes map[kv.Key]kv.Value
-}
+package apitest // import "go.opentelemetry.io/otel/api/apitest"